feat: Expose application MCP via deployments path #1477
