+++ server/src/test/java/com/epam/aidial/core/server/ToolSetApiTest.java
@@ -461,75 +461,6 @@ public class ToolSetApiTest extends ResourceBaseTest {
     }
 
     @Test
-    void testProxyMcpPostCall_WhenToolsetIsAppType() {
-        var response = send(HttpMethod.PUT, "/v1/applications/3CcedGxCx23EwiVbVmscVktScRyf46KypuBQ65miviST/my%20app", null, """
-                {
-                "displayName": "Test Application",
-                "description": "For testing purposes only",
-                "endpoint": "http://<app host>/openai/deployments/test-app/chat/completions",
-                "forwardAuthToken": true
-                }
-                """);
-        verify(response, 200);
-
-        String mcpRequest = """
-                {
-                   "payload": "foo"
-                }
-                """;
-
-        Response resp = send(HttpMethod.POST, "/v1/toolset/applications/3CcedGxCx23EwiVbVmscVktScRyf46KypuBQ65miviST/my%20app/mcp",
-                null, mcpRequest, "Content-type", "application/json");
-        // application doesn't support MCP interface
-        assertEquals(400, resp.status());
-
-        response = send(HttpMethod.PUT, "/v1/applications/3CcedGxCx23EwiVbVmscVktScRyf46KypuBQ65miviST/my%20app", null, """
-                {
-                "display_name": "My App",
-                "display_version": "1.0",
-                "icon_url": "http://apprunner/icon.svg",
-                "description": "My app Description",
-                "applicationTypeSchemaId": "https://mydial.somewhere.com/custom_application_schemas/specific_toolset_type",
-                "applicationProperties": {
-                  "property1": "foo",
-                  "property2": "bar"
-                  }
-                }
-                """);
-        verify(response, 200);
-
-        String mcpResponse = """
-                {
-                  "result": "success"
-                }
-                """;
-        TestWebServer.Handler handler = request -> {
-            try {
-                assertNotNull(request.getHeader(Proxy.HEADER_API_KEY));
-                String body = request.getBody().readString(StandardCharsets.UTF_8);
-                JsonNode tree = ProxyUtil.MAPPER.readTree(body);
-                String basePath = "params._meta.ai_dial_config";
-                JsonNode node = JsonPath.read(tree, basePath + ".property1");
-                assertNotNull(node);
-                assertEquals("foo", node.asText());
-                node = JsonPath.read(tree, basePath + ".property2");
-                assertNotNull(node);
-                assertEquals("bar", node.asText());
-                return new MockResponse().setBody(mcpResponse).setHeader("Content-Type", "application/json");
-            } catch (Throwable e) {
-                return new MockResponse().setResponseCode(500);
-            }
-        };
-        try (TestWebServer ignore = new TestWebServer(9876, handler)) {
-            resp = send(HttpMethod.POST, "/v1/toolset/applications/3CcedGxCx23EwiVbVmscVktScRyf46KypuBQ65miviST/my%20app/mcp",
-                    null, mcpRequest, "Content-type", "application/json");
-
-            assertEquals(200, resp.status());
-            assertEquals(mcpResponse, resp.body());
-        }
-    }
-
-    @Test
     void testProxyMcpPostCall() {
         String mcpRequest = """
                 {
